fix: move MetaStartup to other assembly as .net core 3 bug workaround
