+++ src/Examples/JsonApiDotNetCoreExample/Startups/ClientGeneratedIdsStartup.cs
