+++ src/Examples/JsonApiDotNetCoreExample/Startups/Startup.cs
