--- test/JsonApiDotNetCoreExampleTests/Helpers/Services/IAuthorizationService.cs
@@ -1,7 +1,0 @@
-namespace JsonApiDotNetCoreExampleTests.Services
-{
-    public interface IAuthorizationService
-    {
-        int CurrentUserId { get; set; }
-    }
-}
